+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 ### is.codion.swing.common.model
 - ProgressWorker now supports multiple handlers.
 - ProgressWorker.TaskHandler and related classes added.
+- FilterTableModel.Editor renamed RowEditor.
 ### is.codion.swing.common.ui
 - FilterTableSearchModel.Results.CurrentResult added.
 - DefaultFilterTableSearchModel.DefaultResults.current() now nullable.